Add files via upload
The objective of this dataset i.e. NETFLIX is to analyse the Big Data in apache pig. We have worked on
HDFS cluster and also we have written different queries based on the respective dataset. Different types of
functions are used in designing the queries. The purpose of analysing this dataset is to fetch our requirements
in order to easily search any data which we need. Netflix is a very popular subscription-based streaming
service that allows our members to watch TV shows and movies. Netflix content varies by region and may
change over time. You can watch from a wide variety of award-winning Netflix Originals, TV shows,
movies, documentaries, and more.